Adds legacy docs of early conceptualization.
This is the documentation of the earliest conceptualization of a generic
data management application for trait data, in support of NESCent's
scientists and one of their most common use cases. The conceptualization
was an Informatics group effort led by Hilmar Lapp, and this documentation
was written and developed by Vladimir Gapeyev.

This conceptualization effort dates back to Oct 2008. The results were
not put into design or implementation, but informed design and
implementation of the TraitDB application, which started years later.