Initial search controller logic for summarizing
rows are annotated with the taxon id corresponding to the selected
summarization level.  Then, if summarization was enabled, the rows are
chunked by this id.

Currently, the summarization only selects the first otu in the list -
it does not actually combine the traits/metadata.  But it does reduce
the number of results.